Modify field types for the BaseModels
Changed references to other BaseModels to String because Firebase
returns a String key to be used as the PK when pushing records. Also
added a public constructor for creating BaseModels that have not yet
been associated with other required BaseModels.

e.g. Creating a new Trail is possible without knowing the firebaseId
for the area it is in. However, this value must be bound before it can
be added to the Firebase Database.