Replace FirebaseListener with DatabaseListener
FirebaseListener was too messy and required overriding too many
methods. DatabaseListener is based off the StorageListener and blocks
the thread calling the method in DatabaseProvider until a result is
returned. This allows methods calling it to expect a result.

Also added a disclaimer to remind myself that all these methods should
be accessed on a background thread to prevent locking up the UI thread.