Add initial files for FragmentGuideList
Added SyncGuideTaskLoader to load guides from the Firebase Database.
GuideAdapter is responsible for inflating the layout used for the guide
list items and uses data binding to bind data to the Views.
FragmentGuideList uses an AsyncTaskLoader to load the Guides in a
background thread and then loads them into the Adapter when complete.