Update GpxUtils to pass a MapboxOptions instead of Polyline
In order to center the map on the coordinates of the trail, the trail's
center needs to be calculated. Therefore, the method has been altered
to pass a MapboxOptions Object containing both the coordinates for the
center of the trail and the Polyline representing the trail.

Additionally, a second method for calculating the trail from an
InputStream has been added. This method runs on a separate thread and
takes a Listener interface in the signature to as a callback to inform
the calling Object that the options have been successfully calculated.
This is useful when the user has not selected to save the GPX file to
the internal memory, instead opting for a stream directly from
FirebaseStorage.