Fix the insertRecords method
For inserting Sections, they need to be a child of the FirebaseId of
the Guide that they belong to. Previously, this was checked using an
==. However, since the variable is a String, it needed to be checked
with a equals(). The directory then needed to be modified to append the
FirebaseId to it so that when updateChildren() is called, it finds the
correct directory.