Add custom implementation of MapView
The SmartMapView has a field to keep track of whether it has gone
through onCreate() so that it isn't called multiple times as in the
case of embedding it in RecyclerView. Now, it will only be called once
regardless of how many times RecyclerView re-creates the View. This
will keep it from crashing when it attempts to call onCreate multiple
times per instance.