Implement search function
Once the Place has been selected by the user, the coordinates are used
to use GeoFire to query the location in the database for nearby Guides.
Then the Guides are added to the Adapter to be displayed to the user
and its .gpx file is downloaded to create the Polyline representing the
trail which is then added to the map.