Modify the way track colors are set
Previously, the track colors were set by using a member variable to
hold the position of the track color to generate, but this leads to an
inability to change the colors when Guides are added/removed from the
Adapter. Altered it so that the position of the Guide in the Adapter is
used to set the track color so that when Guides are removed, their
Polyline can be removed and the rest of the track colors can be set to
match the new positions of the Guides in the Adapter.