Implement track highlighting on long press
When a user long-presses a ViewHolder, the swatch will change to yellow
and the associated track will also change to yellow and increase its
width to make it more visible. Long-pressing it again will un-highlight
it. Long pressing another Guide will highlight that track and un-
highlight the previously selected track/Guide.