Modify how MapView is loaded in RecyclerView
Previously, the Style of the map was set in code, but moved the Style
to an XML attribute since it never changes and cleans up unnecessary
code. Also moved the method for setting the camera position to occur
as soon as the MapboxMap is ready so that it doesn't wait for the trail
to be drawn first. This prevents the map from showing the world map and
then snapping to the correct location. The camera zoom is also set in
XML, but will need to be modified when buttons for changing the zoom
level are added.