Modify GuideViewModel to allow it to access offline files
Previously, the getGpx() method always returned a StorageReference.
However, there will be cases where the user is either accessing a Guide
that has been downloaded to offline or when a user is creating a Guide
where the GPX file will instead be stored in local storage and not on
Firebase.