Modify getImeAction() in SectionViewModel
Both Sections types (with/without image) will have IME action of DONE.
Originally, it was planned that sections with text only would have a
NEXT that would trigger the addition of another Section with text only
for more fluid typing. However, this is taking too long to figure out,
so it's being removed until the rest of the app is finished.