Fix issue crashing when adding Author to Firebase Database
When adding authors to the Firebase Database, the Firebase Id is the
ID of the user. This means that the value needs to be set instead of
pushed and updating the child with the model values. Due to
getImageFile() being a public method, the Firebase Database attempts to
retrieve the Uri to add to the database. The BaseModelWithImage
incorrectly attempted to return the File despite there not being a
corresponding Uri. Added a return null in the case of no Uri to work
around this issue. This will never actually be added to the database
because when adding user details, updateChildren() will be used instead.