Add new data model: PlaceModel
PlaceModel is based off the information that Google Places API returns.
It will be used to display items returned by Google Places AutoComplete
as to keep it separated from the Areas that already exist in the
database. Then, when a user clicks on an item, it can be determined
whether it needs to be added to the database, and if so, already
contains the data the database expects.