Add filtering of Trails by name
Modified TrailActivity to pass the Area passed through the Intent to
the SearchTrailViewModel. The ViewModel then queries the Firebase
Database for all Trails associated with the Area. Added a filter method
to the ViewModel. TrailAdapter was modified to use a SortedList and a
Callback to dynamically modify the Adapter List to match the filter.