Add error indication for missing GPX file
Adding this error indication also included tweaking of how the MapView
is loaded. Previously, the MapView would attempt to get an instance of
the MapboxMap every time it was loaded, but sometimes, attempting to
load it would fail.

The source of this bug is still unknown. There doesn't seem to be a
pattern to when it will fail to load. Only that it occurs in
CreateGuideActivity about half the time.

To work around this issue, the method was tweaked so that it does not
attempt to load the MapboxMap every time. Instead, the MapboxMap is set
to a member variable once it is loaded the first time. However, if it
fails to load the first time, the bug will persist.

This bug only affects the view of the GPX file and the error that
shows, but does not affect the usage of the Activity. If the GPX file
has been selected by the user, it will still be able to be published.