Add methods for signing out
Also includes proper loading of the profile after signing in as well as
signing out. When signed in and viewing their own profile, the user
will have have an ActionBar overflow menu option to sign out.

The FAB's scroll behavior is now set programmatically because if it is
set when the user is not viewing their own profile, after scrolling, it
will show the option to create a Guide.