Add Mapbox initialization to MapboxActivity/Fragment
If the app is killed while in background and attempts to re-open the
app to a screen with a MapView, the app crashes because the Mapbox has
not been initialized. Added the Mapbox.getInstance() to the onCreate()
of MapboxActivity/Fragment so that any screen that might have a map,
will ensure that it is initialized and prevent the crash.