Fix loading of map while offline
Due to the way that Mapbox's MapView loads, it will not call finish
loading the map until the maps style is loaded. Since the app uses the
outdoors style rather than the default, it never finishes loading and
only shows a black page where the map should be.

Fixed the issue by setting the map's style before onStart() is called.
This was implemented in SmartMapView's startMapView() method so it is
universal for all MapViews.