Modify how the FAB's visibility is set
Previously, the FAB's visibility was set immediately when the Activity
loaded. However, this leads to a problem if the Author isn't loaded
quickly enough and a user clicks the FAB to start a new Guide. When
this happens, there is not Author loaded to pass to the AreaActivity,
causing a failure when attempting to publish.