Fix loadGpxToMap BindingAdapter in GuideViewModel
Previously, the BindingAdapter was set to not require all parameters.
This caused a problem with other layout attempting to use that
BindingAdapter when they were actually set to a different ViewModel.
Reset the BindingAdapter to require all parameters and added the
XML attribute to any layouts that did require that BindingAdapter.