Add "draft" column to tables
Allows the user to save a Guide as a draft to continue working on at a
later time. This circumvents the need for a second copy of the table to
hold all the draft entries. By querying and filtering for entries that
are true ("1") for the draft column, Guides can be selectively loaded
and shown to the user.