Add handling for draft elements in the database
BaseModel now includes a new variable/methods to indicate whether the
data model is a draft to be completed at a future time.
CreateGuideActivity now sets each element as a draft that needs lacks a
FirebaseId when saving a draft. ContentProviderUtils now includes the
draft column when creating ContentValues for models.