Fix methods for generating ContentValues for models
Previously, the method was directly attempting to add a boolean value
to the database. Since there is no boolean DataType, an Integer was
used with 0 being false and 1 being true. Fixed the methods by adding
the integer value of the boolean for whether the model is a draft to
the ContentValues.