Update GuideViewModel to prefer Uris when loading image/gpx
Since Guide drafts will have a FirebaseId, the presence of one is not a
valid indicator for whether the ViewModel should load the image from
Firebase or from local storage. By checking whether the Uri for an
image is valid, it becomes more reliable because a Guide that is synced
from online cannot have an Uri for its image.