Update layout for CreateGuideActivity
Replaced the FabSpeedDial with a bottom Toolbar. The FAB required two
clicks in order to add anything to the Guide. This was too common an
occurrence for the main function of the Activity. Additionally, the FAB
would obscure the text as the user is writing the Guide.

Accompanying the bottom Toolbar are higher resolution images to be used
for the icons on the Toolbar.

The functions of the Buttons now also scroll the RecyclerView to the
newly added Sections and request focus as to not break the flow of the
user as they type their Guide.