Fix method for loading loading AccountActivity
Previously, when UserFragment was UserActivity, it would simply finish
the Activity and re-load it to prevent Users from seeing a blank
screen if they failed to log in.

Modified the method to start an Activity for result and either load the
main tab if the User failed to log in or load the user's details if
they successfully logged in.