Enable favoriting of Guides in GuideListFragment
The Author is loaded in GuideListFragment and passed to the Adapter and
ViewModels. Each Guide then checks its own FirebaseId to see if it is
within the Author's list of favorites. If it is, then the star icon is
changed to indicate that it is already a favorite. Clicking the star
adds/removes the Guide from the Author's list of favorites on Firebase.