Fix removal of guides when deselected in FavoritesFragment
Previously, clicking the star of a Guide that was already favorite'd
would cause the Guide to be removed from the Author's list of
favorites, but the change wasn't reflected in the Adapter. By using the
fragment tags, the FavoriteFragment is accessed from within the
ViewModel to call the method to remove the Guide from the Adapter.