Fix crash when selecting hero image in CreateGuideActivity
Due to previous changes, selecting an image would cause the
RecyclerView to scroll to the bottom of the List to display the newly
added Section. However, this behavior also carried on when selecting
the hero image to use for the Guide. If there was no Section at the
bottom of the Adapter, it would cause a crash.

Fixed the issue by checking to ensure the FilePicker is selecting an
image for a Section before scrolling the RecyclerView.