Add methods for syncing local database with Firebase
When a user starts saving favorite items offline, but later decides to
create an account, these methods will allow for the User to sync their
locally favorite'd Guides to their online profile.

Additionally, if a user modifies the local database and then later
signs in to their online account, the method will allow for the offline
database to be synced to match their online database, removing any
changes that were done locally.