Modify how images are loaded by ViewModels
Modified the method for retrieving the image source to only return a
Uri. The Uri is then parsed to a StorageReference in the loadImage()
BindingAdapter if the scheme matches.

Since the BindingAdapter is called from the first ViewModel with the
correct parameters, the multiple BindingAdapters for loading images for
each ViewModel caused conflict in which offline images would not
properly load. Removed all duplicate BindingAdapters so that the only
one left is in GuideViewModel.