Remove re-positioning of the MapView
When the ViewHolder is created, the ViewModel containing the MapView
orients itself to fit the Polyline representing the Trail. However,
when it is reloaded, remnant method for positioning the camera would
take over and reposition the camera at a different zoom level.

Removed the remnant code to fix the issue.