Update SearchViewModel delay search
Previously, the method for sending the query to Google Places API was
instantaneous, meaning that it would query several times in the time
that the user was typing their query.

Added a Handler to run the query after 500ms and to cancel any
previously running queries so only one query is sent.