Update search method
Now the search radius is calculated by the distance viewable in the
MapView so that whatever the user can see, is the search radius. Also
added instructions to zoom in if the map is zoomed out further than
level 8 as to prevent searching for too large of a search radius and
returning too many guides.