Fix app crash when entering CreateGuideActivity
When starting a new Guide, there is no image associated with the Guide
and therefore the ViewModel receives null for the image. The method for
loading the image calls for checking the scheme of the Uri, which
causes the crash as the Uri is null.

Fixed the issue by checking and skipping the process if the Uri is null.