Update Rating data model to include dateAdded
When the data is pulled from Firebase Database, they are ordered by the
userId of the rater. In order to sort by date, a Field needs to be
added to hold the value.

When uploading the Rating, it uses the server's timestamp as to prevent
errors in adding date based on the device's time.