Fix Rating data model getDateAdded() method
Due to the way that Firebase converts Objects to be saved in their
database, the getDateMethod needed to be made flexible to handle either
return the value to get the server timestamp or the value of the date
that already existed.

e.g. Rating1: dateAdded = 1 & Rating2: dateAdded = 2
If the user updates Rating1, then the method needs to return the the
original value for dateAdded for Rating2 and the value to get the
timestamp for Rating1 to prevent Rating2 from updating its dateAdded
value to the server timestamp.