Add method to GuideDetailsFragment to load Guide from Firebase
When a user adds a rating for the first time, if they return to
GuideListFragment and then click on the same Guide, the Guide will not
be updated because it was using the Guide passed from
GuideListFragment.

By re-loading the Guide from Firebase, it ensures that if the same
circumstances happen, the user's review will be visible.