fix: Fix SavedGuidesFragment loading un-published Guides
The CursorLoader didn't have the correct selection and would load all
Guides that had an image Uri indicating that they had been saved to
local storage, not taking into account that Guides that are saved as a
draft also have image Uris.

Fixed the issue by setting the selection to also check the dateAdded
column to ensure that the Guide is a saved published Guide and not a
draft.