feat: Add DataBinding to PublishActivity
Includes the addition of PublishViewModel for handling updating Views
for the Activity. The Activity now displays an indeterminate
ProgressBar with a TextView indicating the the upload progress of the
Guide.

The ProgressBar was not set to determinate to show the progress of each
upload because Firebase did not give enough updates for each upload. It
would only update the progress at 0 bytes transferred and file fully
uploaded, regardless of speed of upload.