fix: Fix deletion of draft entries after publishing
When the user publishes a Guide, the draft is deleted from the
database. However, the method previously attempted to delete the
entries using the FirebaseIds that were given to the models upon
publishing. This mean that their entries were not found in the database
and would be retained.

Fixed the issue by creating dummy models with the original FirebaseIds
so they could be properly located and deleted from the database.