fix: Fix Adapter not updating when Guides are deleted
In SavedGuidesFragment, if the user clicks on a Guide and then deletes
it from local storage and returns to SavedGuidesFragment, the guide
would still be there because the Adapter did not update to reflect the
changes.

Now when the Loader returns data, it clears the Adapter so that is can
properly update and display the Guides that still exist in local
storage.