fix: Fix crash when attempting to save a guide
When users are saving very short guides with very few files to
download, it caused a crash. This occurred because the order in which
the GPX file for the Guide was set happened after attempting to save
the Guide. This only occurs on guides with few downloads because
otherwise the other downloads occurring would give the app enough time
to properly set the GPX file to the guide.

Fixed the issue by correcting the order of how the GPX file is set to
occur before the guide is saved.