fix: Resolve issue of Ratings not being properly updated
The Rating passed to RatingViewModel is modified when the user
interacts with it, causing the "previousRating" variable to be the same
as the new rating given by the user. Due to the way that the Ratings
modified the Author of the Guide's score and the Guide's rating, it
would result in a net change of zero.

Fixed the issue by creating a duplicate of the Rating that is modified
when the User interacts with it. This allows the originalRating to
remain the same so that when it is updated, it still has the correct
"previousRating."