fix: Resolve app crash when loading reviews while logged in
The issue arose when a user was logged in to their account, had not
reviewed a guide yet, and another user had already reviewed the guide.
The occurred because the new Rating created for the logged in user did
not have a FirebaseID assigned to it yet. When adding the previous
reviews to the Guide, it would check it against the FirebaseID of the
new Rating to see if the Rating had already been added to the Adapter,
causing a NPE.

Resolved the issue, by checking to ensure the FirebaseID of the model
was set before checking it against the newly added model.