feat: Modify rating sections heading for guides
Moved the heading from about the author section to the top of the
rating section. This makes more sense as the heading never really
belonged with the author section and was only done out of convenience.

This also gives the added benefit that the section heading does not
show when there are no reviews to be shown.