feat: Add DataCache class
DataCache is used to store data models to a memory cache so that they
can be retrieved from other Activities/Fragments. This reduces the
number of network calls that need to be made as the data will already
be stored in cache and will not need to be downloaded every time.

This also gives the advantage of allowing user-initiated changes be
immediately visible when returning from a different Activity.

e.g. The user rates the guide, changing its rating and then returns to
GuideListFragment. The change in rating is immediately seen as the data
was changed in the cached Object which is then used to populate the
Adapter.