fix: Resolve crash when attempting to update a rating
The issue occurred because when the Ratings are pulled from Firebase,
the guideId and guideAuthorId were not properly set for the Rating.
When the user attempted to update the rating, those parameters were
missing, causing a crash from reading a null Field.

Fixed the issue by properly setting the guideId and guideAuthorId when
the Ratings are retrieved from the database.