fix: Resolve crash when opening a Guide from FavoritesFragment
Prior to loading the Guide from local database, the method for checking
whether a Guide existed in the database didn't take into account
whether it was just a favorite'd entry or if its info was saved in the
database.

Fixed the issue by adding a method specifically for checking whether a
Guide is saved in the database vs just favorite'd.