feat: Add saving of RecyclerView scroll state
Previously, on configuration change, the data downloaded from Firebase
would be lost, and the data would need to be queried again. This made
it impossible to restore the scroll state as the RecyclerView could not
scroll to a position that had not been created yet.

By using the DataCache and saving the FirebaseIds of the guides that
were previously loaded in onSaveInstanceState, the guides and scroll
state are able to be restored on configuration change.