fix: Fix incorrect icon showing in GuideDetailsFragment
The icon in the ActionBar was showing the icon for deleting a cached
Guide when the guide was not cached. This issue arose due to the method
for setting the icon utilizing the wrong method for checking whether
the guide was actually cached in the database.

Fixed the error by calling the correct method for checking whether the
guide was cached.