fix: Resolve issue with infinite loading icon
In GuideDetailsFragment, when the user selects a guide that is saved
to local storage, the loading icon in the ActionBar never stops even
after the guide is loaded. This issue occurred because when loading
from local database, the method to stop the loading icon is never
called.

Fixed the issue by calling the proper method to stop the loading icon
after all data has been loaded from database.