fix: Resolve crash when attempting to publish a Guide
Some kind of error in reading a Parcelable from the Intent caused a
crash in unmarshalling the Parcel. Not sure how this came about
considering ALL these Objects are Parcelized and read back in other
Activities exactly the same way.

Worked around the issue by saving the data to the DataCache and reading
it back in PublishActivity.