fix: Crash when attempting to write a new Guide
When opening the AreaActivity, there was a crash due to the changes in
the GuideViewModel altering the BindingAdapter to not require all
values. This caused the Mapbox MapView to call the incorrect
BindingAdapter.

Fixed the issue by altering the BindingAdapter to check for null values
prior to using the variable.