fix: Resolve issue restoring state in GuideDetailsActivity
In GuideDetailsActivity, because the images in the guide do not have
fixed aspect ratios, when the user rotates the screen, the image
heights are initially set to 0 (wrap_content). This meant that although
the scroll state was restored, it would be in the incorrect position
after the heights of the ImageViews were properly calculated.

Fixed the issue by adding a custom class AspectRatioImageView capable
of setting the height of the ImageView based on an aspect ratio. This
aspect ratio is calculated the first time the images are loaded and
stored in the Section data model. When the user rotates the screen,
the ImageView's heights are calculated by the ratio set by the
accompanying Section.