feat: Move database operations to background thread
Previously, when the user loads their own profile in UserFragment, the
operation to clean the profile would be run every time and on the UI
thread, causing a slight delay while loading the profile.

Modified the method to only clean up the database entries when the user
logs in for the first time. Also utilized a new Thread to complete the
operation as to avoid using the UI thread.