fix: Resolve crash when opening a Guide from widget
Opening a Guide from the widget forces the information to be downloaded
from Firebase because there is no guide cached. When attempting to
download the information, the author's information is attempted to be
retrieved before the Guide's authorId has been populated.

Fixed the issue by running the method for retrieving author information
after guide information is completely downloaded.