fix: Resolve crash #2 when opening a Guide from widget
When loading GuideDetailsMapFragment from the widget, the Guide is not
populated with the required information for starting the MapView. When
attempting to load the GPX information, it crashed because that info
did not yet exist.

Resolved the issue by adding a separate method for loading the Guide
information from Firebase and then initializing the MapView after it
had been downloaded.