fix: Resolve crash #3 when opening a Guide from widget
When loading GuideDetailsMapFragment from the widget, the
ViewDataBinding for the layout automatically attempted to start the
MapView. However, because no ViewModel had been loaded to it containing
the required Activity/Fragment for starting the MapView.

Fixed the issue by adding a check to ensure that the Activity is not
null prior to starting the MapView.