refactor: Split responsibility for toggling favorites
The methods for toggling the favorite status of a Guide in
ContentProviderUtils and FirebaseProviderUtils contained code for
toggling the favorite status in the Guide object, violating the single
responsibility principle.

Refactored the code for toggling the favorite status of the Guide
Object to the click response of the favorite button.