feat: Add favorites to local database when logged int
Previously, when a user was logged in, the favorite status of guides
would only be altered in their online account. This meant that their
favorites would not show in the widget as it is only allowed to pull
information from a local source.

Added toggling the favorite status of
users who are logged in to a local database to allow them to also use
the widget.