fix: Resolve crash when loading a Guide with no Ratings
When a user is logged in and a Guide has no ratings, the
GuideDetailsAdapter attempted to generate the hashcode from a null
FirebaseId because the new Rating did not have a FirebaseId assigned to
it yet.

Fixed the issue by checking to ensure that the FirebaseId is not null
before trying to generate a hashcode for it.