feat: Semi-enable D-pad navigation in SearchFragment
Due to the way that MapboxMap requires focus in order to change the
camera, there is no way to avoid giving it focus. Once it has focus, it
uses D-pad navigation to pan the camera. There doesn't seem to be a way
to disable those controls in either XML or Java.