feat: Add attribution when searching Google Places API
When the initial query goes against the Firebase Database, there is no
need to show the Google logo in the attribution. Then, if the user
cannot find the park/area that they are searching for and choose to
search Google Places API, the Google logo shows in the attribution bar.