feat: Add NewAreaAdapter for SearchAreaActivity
This Adapter will eventually supersede and replace AreaAdapter. It is
more precise in handling the items in the Adapter due to the use of a
SortedList. It also implements the new generic ClickHandler that will
replace all ClickHandlers in Adapters.

It also offers better handling of the extra list item that is not part
of the data to be displayed.