fix: Fix method for setting items in NewAreaAdapter
Previously, the method incorrectly removed items from the Adapter prior
to setting the new list for the Adapter, causing items to only be
added.

Also fixed the signature of the method so that it takes a List of items
that extend AreaAdapterSortable instead of a List of
AreaAdapterSortable so that the items returned from the search queries
can be properly added to the Adapter.