fix: Resolve issues trying to display Markers
Markers were not properly added and removed because when a Marker was
created for a Trail, it was created using the MarkerOptions used to set
the location of the Marker. This Marker object was not the same Marker
added to the MapboxMap.

Fixed this issue by setting the newly added Marker to the MapboxMap as
the Marker to be stored/removed.

When the Adapter was reset, the Markers were not re-added to the
MapboxMap.

Fixed the issue by re-adding the Markers when the Adapter was reset.