fix: Resolve issue selecting Area causing query to disappear
In AreaActivity, when selecting to show the Area's location, it would
clear the user's query. This would also happen when the user selected
the search box again after finishing viewing an area's location.

The issue occurred because the focus for setting the focus also set the
selected Area to null. This reset the query.

Fixed the issue by setting the Area to null only if the Area is not
already null.