feat: Update NewAreaAdapter to prioritize Areas over Places
When a list of Places is returned from Google Places API, the names of
the locations are checked against the Areas already in the Adapter. If
the name matches a place already in the Adapter, then the the item in
the Adapter remains and the one returned from Google Places API is not
added.

This is an attempt to keep the database of Areas clean and prevent
duplicates where possible.