feat: Optimize GuideDetailsAdapter loading of ViewHolders
Previously, the RecyclerView would recycle the ViewHolder for the
guide details containing the MapView. This would cause stutter while
scrolling the RecyclerView as it is an expensive process to re-load the
MapView.

The Adapter was optimized by setting the ViewHolder containing the map
to not recycle and reload its original contents when scrolling.