feat: Modify how Ratings are stored in Firebase Database
Instead of storing the Ratings under the FirebaseId of each separate
Guide, they are instead stored under their individual keys. This will
allow all the ratings of a guide or author to be easily queried and
returned.

Unfortunately, this also means that to pull a Rating specific to a
guide and author, it will be required to pull all the Ratings for an
author and then filter for the one matching the specific guide.