fix: Fix crash when loading guide details
When loading the ratings for the guide, there was a mistake where the
Rating getGuideId() was annotated with @Exclude so the value wasn't
added to the Rating when it was downloaded from Firebase Database.

This caused a NPE as it tried to compare a null guideId.

Fixed the issue by removed the @Exclude annotation.