feat: Update method for updating guide/author score
Instead of having the Transaction be re-run whenever it fails, it is
cleaner to automatically re-run the Transaction if Firebase fails to
return a valid object for the guide or author.

Still not sure what is causing the issue as subsequent runs of the same
method with the same parameters will eventually return valid data.