feat: Rename app to Sherpa
Due to some weird glitch, the git repo became bare upon renaming the
app. Resolved the issue by cloning the project to another folder and
then copying the .git folder back to the working directory.

This is causing a lot o files to appear as if they are being removed
and re-added when in fact, they have just been moved.