fix: Fix favorite button in UserFragment
Previously, the logged in user was not synced to the Guides loaded into
the Adapter. This meant that the favorite status of the Guides were not
synced to the Firebase Database, only the local database, causing
discrepancies to arise when the user favorite'd a Guide from
UserFragment.

Fixed the issue by loading the logged in user and setting it so that
the Guides can check their favorite status against the Author's
favorites and set it appropriately in Firebase when necessary.

Resolves: #4