fix: Fix issue with SearchFragment save state
When SearchFragment saves its information prior to a state change, the
Polylines were restored when the Guides were restored. However, when a
GeoFire Query was run, they would register new Polylines instead of
using the ones that were already on the map. This caused them to remain
after the user had moved the camera outside of the search radius of the
GeoFire Query.

Additionally, because the GeoFire Query was reset, the keys of the
guides within the search radius were not registered as having
"entered." This meant that if the user moved the camera outside of the
search radius, they would not be removed from the Adapter.

Fixed the first issue by adding a check to make sure that the Polyline
hadn't already been added to the map before attempting to add a new one.

Fixed the second issue by triggered a GeoFire Query upon restoring the
savedInstanceState so that the keys of the guides were properly
registered.

Resolves: #3