fix: Fix crash when attempting to save guide draft
When the Author is passed from one Activity to the next, the Parcelable
method did not also pass the username. This caused a NPE when
attempting to convert the author's username to lowercase when adding it
to the database.

Fixed the issue by adding the username to the Parcelable method.