feat: Add method for rotating images on upload
Because how rotation data is stored differently by different cameras,
some rotation data is hardcoded in the image dimensions and others are
stored in the EXIF data of the image, some images are not properly
rotated when it is uploaded to Firebase Storage.

Added a method for automatically rotating the images that have the data
in the EXIF file.