fix: Fix crash when saving a Guide
When a user deletes a saved guide and then immediately attempts to save
the guide again, the app crashes because the method for converting a
Cursor to an Author object did not also add the username, leading to a
NPE when attempting to add the Author back to the database.

Fixed the issue by adding the username when converting from Cursor to
Author.