fix: Fix crash when loading MessageAdapter items
In the getItemViewType() method, there was a crash when it attempted to
load the Message corresponding to the position of the item. This
occurred because the method attempted to load a Message for a position
that did not exist in the SortedList in the case of loading the bottom
spacer view.

Fixed the issue by checking whether the position corresponded to the
spacer view prior to attempting to load the Message for the
corresponding position.

Also fixed the issue with adding duplicate messages because there was
no check to ensure the item did not already exist in the Adapter.