feat: Update notification Uri for Messages
The InexactContentUri method for querying the table had the
notification Uri set to the specific Uri for the query. However, this
meant that new messages that were added using the content Uri of the
table would not notify the Cursor of changes.

Set the annotation to change the notification Uri to the content Uri
instead of the Uri used to query for messages.