feat: Update ChatFragment retrieval of Chats
Updated the Fragment so that Chats are updated in real-time and made it
so that the users involved in the chat are added to the local database
before the Chat is added to the Adapter. This allows for the author's
information to be properly loaded into the Adapter as well.