feat: Update Chat data model
Changed field members to activeMembers and created a new field for
allMembers.

The activeMembers will be used to store users that are still in the
chat (i.e. they haven't left the chat). allMembers stores the authorId
of every member that has participated in the chat for the purpose of
being able to load their names and images.