fix: Fix click response in NewChatActivity
The function for setting an Author as selected worked properly, but
because the function for loading the author's image happens in a
background thread, it overrode the image used to show a selected item.

Fixed the issue by adding a check to ensure that the item is not
selected prior to loading the image from FirebaseStorage.