fix: Fix messaging multiline
Due to the way that IME Options are set up, when setting the inputType
to textMultiLine, it overrides the IME Option.

In order to get both a multiline EditText and the IME Option as
ACTION_SEND, the IME Option had to be set in Java code after the layout
had already been determined to be multi-line.