fix: Fix read Chats still showing as unread
In MessageFragment, the method for adding the updated Chat was not run
on the first load when the user first enters the Chat. It was only run
for subsequent updates to the Chat.

Fixed the issue, but setting the update to cause the local database to
be refreshed with the new information every time.