fix: Fix cached Objects being GC'd between Activities
Due to the use of a SoftReference to hold the reference to cached data,
sometimes the data is GC'd when user is moving between Activities. This
causes a NPE because the Activity is expecting a cached Object that no
longer exists.

Fixed the issue by adding a lock in when storing the Object. After
storing the Object, a new Thread is created with a hard-reference to
the stored Object for 100ms. This gives it enough time to survive a
change in Activities without preventing it from being GC'd later on.