fix: Fix crash when receiving messages
Crash occurs because more than one ValueEventListener is added to the
DatabaseReference listening for new messages. When the Fragment is
ended, only one ValueEventListener to removed. Then when the user
receives another message while not in the Fragment, it calls for a
function that requires the Context passed from MessageFragment which
has been destroyed, resulting in an NPE and crash.

Fixed the issue by switching to SmartValueEventListener which ensures
that only one ValueEventListener is attached to a DatabaseReference.